+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.27] - 2023-01-17
+### Added
+* GameObjectExtensions class to Extensions directory
+### Changed
+* TransformExtensions's parameter name to match its type
+### Fixed
+* warning within CameraFacer by renaming camera field
+
 ## [0.0.26] - 2022-11-29
 ### Added
 * ListExtensions function to get the ToString of all elements in list